requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.4.0
aiohttp==3.10.6
aiosignal==1.3.1
alabaster==1.0.0
alembic==1.13.3
annotated-types==0.7.0
anomalib==1.1.1
antlr4-python3-runtime==4.9.3
anyio==4.6.0
argcomplete==3.5.0
asttokens==2.4.1
attrs==24.2.0
awscli==1.35.0
azure-containerregistry==1.2.0
azure-core==1.31.0
azure-identity==1.18.0
azure-storage-blob==12.23.1
babel==2.16.0
beautifulsoup4==4.12.3
bleach==6.1.0
boto3==1.35.25
botocore==1.35.34
Brotli==1.1.0
build==1.2.2
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
colorlog==6.8.2
comm==0.2.2
contourpy==1.3.0
cryptography==43.0.1
cycler==0.12.1
dacite==1.7.0
datasets==3.0.0
debugpy==1.8.6
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.14
descartes==1.1.0
dill==0.3.8
dnspython==2.6.1
docker-pycreds==0.4.0
docstring_parser==0.16
docutils==0.16
durationpy==0.9
einops==0.8.0
executing==2.1.0
fastjsonschema==2.20.0
fiftyone==1.0.0
fiftyone-brain==0.17.0
fiftyone-desktop==0.34.1
fiftyone_db==1.1.6
filelock==3.16.1
fire==0.6.0
fonttools==4.54.1
FrEIA==0.2
frozenlist==1.4.1
fsspec==2024.6.1
ftfy==6.2.3
future==1.0.0
gitdb==4.0.11
GitPython==3.1.43
glob2==0.7
google-api-core==2.20.0
google-auth==2.35.0
google-cloud-aiplatform==1.69.0
google-cloud-artifact-registry==1.11.5
google-cloud-bigquery==3.26.0
google-cloud-compute==1.19.2
google-cloud-core==2.4.1
google-cloud-resource-manager==1.12.5
google-cloud-storage==2.18.2
google-crc32c==1.6.0
google-resumable-media==2.7.2
googleapis-common-protos==1.65.0
graphql-core==3.2.4
greenlet==3.1.1
grpc-google-iam-v1==0.13.1
grpcio==1.66.2
grpcio-status==1.66.2
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.5
httpx==0.27.2
huggingface-hub==0.25.1
humanize==4.10.0
Hypercorn==0.17.3
hyperframe==6.0.1
idna==3.10
imageio==2.35.1
imagesize==1.4.1
imgaug==0.4.0
importlib_resources==6.4.5
inflate64==1.0.0
iniconfig==2.0.0
ipykernel==6.29.5
ipython==8.27.0
ipywidgets==8.1.5
iso8601==2.1.0
isodate==0.6.1
jedi==0.19.1
Jinja2==3.1.4
jmespath==1.0.1
joblib==1.4.2
jsonargparse==4.33.1
jsonlines==4.0.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyterlab_pygments==0.3.0
jupyterlab_widgets==3.0.13
kaleido==0.2.1
kiwisolver==1.4.7
kornia==0.7.3
kornia_rs==0.1.5
kubernetes==31.0.0
kubernetes_asyncio==31.1.0
lazy_loader==0.4
lightning==2.4.0
lightning-utilities==0.11.7
llvmlite==0.43.0
Mako==1.3.5
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.2
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.0.2
mongoengine==0.24.2
motor==3.5.3
mpmath==1.3.0
msal==1.31.0
msal-extensions==1.2.0
multidict==6.1.0
multiprocess==0.70.16
multivolumefile==0.2.3
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.3
numba==0.60.0
numpy==1.26.4
# Editable install with no version control (nuscenes-devkit==1.1.9)
-e /home/dbogdoll/mcity_data_engine/data_loader/nuscenes-devkit
nvidia-cublas-cu12==12.4.2.65
nvidia-cuda-cupti-cu12==12.4.99
nvidia-cuda-nvrtc-cu12==12.4.99
nvidia-cuda-runtime-cu12==12.4.99
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.0.44
nvidia-curand-cu12==10.3.5.119
nvidia-cusolver-cu12==11.6.0.99
nvidia-cusparse-cu12==12.3.0.142
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.4.99
nvidia-nvtx-cu12==12.4.99
oauthlib==3.2.2
omegaconf==2.3.0
onnx==1.17.0
open_clip_torch==2.26.1
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
openvino==2024.4.0
openvino-telemetry==2024.1.0
optuna==4.0.0
packaging==24.1
pandas==2.2.3
pandocfilters==1.5.1
parso==0.8.4
pdoc==14.7.0
pexpect==4.9.0
pillow==10.4.0
platformdirs==4.3.6
plotly==5.24.1
pluggy==1.5.0
portalocker==2.10.1
pprintpp==0.4.0
priority==2.0.0
prompt_toolkit==3.0.48
proto-plus==1.24.0
protobuf==5.28.2
psutil==6.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
py7zr==0.22.0
pyarrow==17.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pybcj==1.0.2
pycocotools==2.0.8
pycparser==2.22
pycryptodomex==3.20.0
pydantic==2.9.2
pydantic_core==2.23.4
pydash==8.0.3
Pygments==2.18.0
PyJWT==2.9.0
pymongo==4.8.0
pynndescent==0.5.13
pyparsing==3.1.4
pyppmd==1.1.0
pyproject_hooks==1.2.0
pyquaternion==0.9.9
pytest==8.3.3
python-dateutil==2.9.0.post0
pytorch-lightning==2.4.0
pytz==2024.2
PyYAML==6.0.2
pyzmq==26.2.0
pyzstd==0.16.1
rarfile==4.2
referencing==0.35.1
regex==2024.9.11
requests==2.32.3
requests-oauthlib==2.0.0
retrying==1.3.4
rich==13.9.1
rich-argparse==1.5.2
rpds-py==0.20.0
rsa==4.7.2
s3transfer==0.10.2
safetensors==0.4.5
scikit-image==0.24.0
scikit-learn==1.5.2
scipy==1.14.1
sentry-sdk==2.15.0
setproctitle==1.3.3
setuptools==75.1.0
shapely==2.0.6
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
snowballstemmer==2.2.0
sortedcontainers==2.4.0
soupsieve==2.6
sphinxcontrib-applehelp==2.0.0
sphinxcontrib-devhelp==2.0.0
sphinxcontrib-htmlhelp==2.1.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==2.0.0
sphinxcontrib-serializinghtml==2.0.0
SQLAlchemy==2.0.35
sse-starlette==0.10.3
sseclient-py==1.8.0
stack-data==0.6.3
starlette==0.39.0
strawberry-graphql==0.138.1
sympy==1.12
tabulate==0.9.0
tenacity==9.0.0
termcolor==2.4.0
texttable==1.7.0
threadpoolctl==3.5.0
tifffile==2024.9.20
timm==1.0.9
tinycss2==1.3.0
tokenizers==0.20.0
tomli==2.0.2
torch==2.4.1+cu124
torchaudio==2.4.1+cu124
torchmetrics==1.4.2
torchvision==0.19.1+cu124
tornado==6.4.1
tqdm==4.66.5
traitlets==5.14.3
transformers==4.45.0
triton==3.0.0
typeshed_client==2.7.0
typing_extensions==4.12.2
tzdata==2024.2
tzlocal==5.2
umap-learn==0.5.6
universal-analytics-python3==1.1.1
urllib3==2.2.3
voxel51-eta==0.13.0
wandb==0.18.3
wcwidth==0.2.13
webencodings==0.5.1
websocket-client==1.8.0
wheel==0.44.0
widgetsnbextension==4.0.13
wrapt==1.16.0
wsproto==1.2.0
xformers==0.0.28.post1
xmltodict==0.13.0
xxhash==3.5.0
yarl==1.12.1
